Clean up api
