+++ src/api/meetings.js
@@ -1,0 +1,50 @@
+import agent from 'superagent'
+import moment from 'moment'
+
+import { apiBaseUrl } from './utils'
+
+export const createMeeting = (token, meeting, room) => agent
+  .post(`${apiBaseUrl}/room/${room.email}/meeting`)
+  .set('x-access-token', token)
+  .send({
+    title: meeting.title,
+    start: meeting.start,
+    end: meeting.end,
+  })
+  .then(response => response.body)
+
+export const fetchMeetings = (token, startDate, endDate) => {
+  let start = startDate
+  let end = endDate
+
+  if (!startDate) {
+    start = moment().startOf('day').format('YYYY-MM-DD')
+  }
+
+  if (!endDate) {
+    end = moment(start).add(1, 'day').format('YYYY-MM-DD')
+  }
+
+  return agent
+    .get(`${apiBaseUrl}/rooms/nyc/meetings?start=${start}&end=${end}`)
+    .set('x-access-token', token)
+    .then(response => response.body)
+}
+
+export const editMeeting = (token, meeting, roomEmail) => {
+  console.log('edit meeting', meeting)
+  return agent
+    .put(`${apiBaseUrl}/room/${roomEmail}/meeting/${meeting.id}`)
+    .set('x-access-token', token)
+    .send({
+      userMeetingId: meeting.userMeetingId,
+      title: meeting.title,
+      start: meeting.start,
+      end: meeting.end,
+    })
+    .then(response => response.body)
+}
+
+export const cancelMeeting = (token, meetingId, roomEmail) => agent
+  .delete(`${apiBaseUrl}/room/${roomEmail}/meeting/${meetingId}`)
+  .then(message => message)
