+++ src/api/utils.js
@@ -1,5 +1,4 @@
 /* eslint-disable no-underscore-dangle */
-
 const configParam = (key, defaultVal) => (window.__CONFIG && window.__CONFIG[key]) || defaultVal
 
-export default configParam
+export const apiBaseUrl = configParam('apiBaseUrl', 'http://localhost:8888')
